Set up response formatter modules, update tests
